+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.1] - Unreleased
+## [0.9.2] - 2025-05-15
+
+### Fixed
+
+- Turtle encoding of IRIs that contain url escape (e.g. %20) must not use prefixes
+
+## [0.9.1] - 2025-05-15
 
 ### Changed